[FIX] hr_holidays: display warning on wrong accrual balance
__Description of the issue:__
The way that accrual works now is that future leaves are
not taken into account for the balance both for the dashboard
display and upon taking leaves.
It is then possible to create situation that would lead to issues
by taking leaves in a certain order; those issue are avoided with
a cron that automatically cancels those issuing leaves.
However, if leaves are causing issues and might be cancelled
by the cron, a warning should appear on the dashboard of the
employee to indicate that issue.

To reproduce the issue:
- Have any kind of accrual plan; the demo seniority for example
- Have a time off type that requires allocation
- Create an allocation for any employee for that accrual plan starting now
and manually grant 1 day
- Go on that employee's dashboard
- Create a leave in a few days:
it should be possible because one day is available for the employee
- Create another one the day before:
the system allows it but it  is a discrepancy in the balance
- Look on the balance for the time off type: no warning is displayed

__Expected behaviour:__
There should be a warning after the creation of the second

__Description of the bugfix:__
This commit fixes the way the excess is detected to make it accurate.
Additionally, the different warning messages were previously made incompatible
with one another though they're not linked; this commit allows multiple
warning messages if multiple are to be displayed.

task-3859558

closes odoo/odoo#161107

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>